refactor(utilities/ast): rename pattern-vitest-mock.ts to pattern-test-mock.ts
