+++ packages/utilities/ast/src/index.ts
@@ -23,6 +23,6 @@ export * from "./node-traverse";
 export type * from "./node-types";
 export * from "./node-unwrap";
 export * from "./pattern-process-env-node-env";
-export * from "./pattern-vitest-mock";
+export * from "./pattern-test-mock";
 export * from "./property-name";
 export * from "./property-traverse";
